Refactorization of viewer
